pyproject.toml
[project]
name = "stocksensei"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.15",
    "langchain-openai>=1.1.14",
    "langchain-anthropic>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-groq>=0.2.0",
    "langchain-ollama>=0.3.0",
    "langgraph>=0.2.0",
    "openai>=1.0.0",
    "prompt-toolkit>=3.0.0",
    "python-dotenv>=1.2.2",
    "rich>=13.0.0",
    "yfinance>=1.3.0",
]

[project.scripts]
stocksensei = "main:main"

[tool.setuptools]
py-modules = ["main", "agent", "tools", "config", "utils", "ui_blocks", "command_prompt"]
